Made the requested changes & rename
